✏️ Modify: fk 표기 수정
