+++ fanmon/src/main/java/com/example/fanmon/domain/board/entity/Groupfanboard.java
@@ -1,4 +1,4 @@
-package com.example.fanmon.board.entity;
+package com.example.fanmon.domain.board.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
@@ -7,19 +7,20 @@ import java.util.UUID;
 
 @Entity
 @Data
-@Table(name="groupartistboard")
-public class Groupartistboard {
+@Table(name="groupfanboard")
+public class Groupfanboard {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "fanboarduuid", updatable = false, nullable = false)
-    private UUID groupartistboarduuid;
+    private UUID groupfanboarduuid;
 
 //    @OneToOne
 //    @JoinColumn(name="groupuuid", nullable = false)
-//    private Group groupuuid;
+//    private Group group;
 
     @OneToOne
-    @JoinColumn(name="artistboarduuid", nullable = false)
-    private Artistboard artistboarduuid;
+    @JoinColumn(name="fanboarduuid", nullable = false)
+    private Fanboard fanboard;
+
 }
